merge: sync with origin/develop and resolve conflicts
Resolve 6 merge conflicts from stash pop after pulling origin/develop:
- tests/test_pipeline_smoke.py: keep remote (new regression tests)
- tests/test_numba_warmup.py: keep remote (new float32 tests)
- tests/test_grain.py: keep remote (new RNG/immutability tests)
- src/spektrafilm/utils/io.py: keep remote (float16 overflow warnings)
- src/spektrafilm/generators/spectral_generator.cpp: keep local (RDom bounds fix)
- README.md: keep local (language consistency)

Also includes prior local work on HDR, gain maps, Halide backend,
spectral upsampling, and extended test coverage.

+++ src/spektrafilm/utils/spectral_upsampling.py
@@ -99,7 +99,7 @@ def compute_lut_spectra(lut_size=128, smooth_steps=1, lut_coeffs_filename='hanat
     lut_coeffs = _load_coeffs_lut(lut_coeffs_filename)
     coeffs = _fetch_coeffs(tc, lut_coeffs)
     lut_spectra = _compute_spectra_from_coeffs(coeffs, smooth_steps=smooth_steps)
-    lut_spectra = np.array(lut_spectra, dtype=np.half)
+    lut_spectra = np.array(lut_spectra, dtype=np.float32)
     return lut_spectra
 
 def _load_hanatos2025_spectra_lut(filename='irradiance_xy_tc.npy'):
